API: Split to multiple controllers, removed / from routes
